Decompose dsl_serializer into 5 focused files
Extract type_helpers, rule_serializer, behavior_serializer,
aggregate_serializer as includable modules. Fixes DuplicateMethodCall
and TooManyStatements.

Phase 3 of HEC-493.
